@stratusjs/swiper 0.1.0 upgraded to typescript fixed random module loading issues
