+++ package.json
@@ -60,6 +60,7 @@
     "@types/jquery": "^3.3.31",
     "@types/lodash": "^4.14.138",
     "@types/node": "^12.6.8",
+    "@types/swiper": "^4.4.6",
     "@types/systemjs": "^0.20.6",
     "@typescript-eslint/typescript-estree": "^1.12.0",
     "angular": "^1.7.8",
